Amélioration lisibilité POO
