+++ Livrable1/bin/Debug/net8.0/state.json
@@ -1,38 +1,38 @@
 [
   {
     "Appellation": "Sauvegarde_1",
-    "HorodatageDerniereAction": "2025-02-05T17:28:06.7942127+01:00",
+    "HorodatageDerniereAction": "2025-02-07T16:18:13.7085365+01:00",
     "Etat": "END",
     "NombreTotalFichiers": 100,
     "TailleTotaleFichiersMB": 5000,
     "Progression": 100,
     "NombreFichiersRestants": 0,
     "TailleFichiersRestantsMB": 0,
-    "FichierSource": "/mnt/data/source/file57.txt",
+    "FichierSource": "/mnt/data/source/file99.txt",
     "FichierDestination": "/mnt/data/destination/file55.txt"
   },
   {
     "Appellation": "Sauvegarde_2",
-    "HorodatageDerniereAction": "2025-02-05T17:28:06.7942174+01:00",
+    "HorodatageDerniereAction": "2025-02-07T16:18:13.708555+01:00",
     "Etat": "END",
     "NombreTotalFichiers": 100,
     "TailleTotaleFichiersMB": 5000,
     "Progression": 100,
     "NombreFichiersRestants": 0,
     "TailleFichiersRestantsMB": 0,
-    "FichierSource": "/mnt/data/source/file58.txt",
-    "FichierDestination": "/mnt/data/destination/file16.txt"
+    "FichierSource": "/mnt/data/source/file20.txt",
+    "FichierDestination": "/mnt/data/destination/file1.txt"
   },
   {
     "Appellation": "Sauvegarde_3",
-    "HorodatageDerniereAction": "2025-02-05T17:28:06.7942192+01:00",
+    "HorodatageDerniereAction": "2025-02-07T16:18:13.708557+01:00",
     "Etat": "END",
     "NombreTotalFichiers": 100,
     "TailleTotaleFichiersMB": 5000,
     "Progression": 100,
     "NombreFichiersRestants": 0,
     "TailleFichiersRestantsMB": 0,
-    "FichierSource": "/mnt/data/source/file19.txt",
-    "FichierDestination": "/mnt/data/destination/file6.txt"
+    "FichierSource": "/mnt/data/source/file69.txt",
+    "FichierDestination": "/mnt/data/destination/file18.txt"
   }
 ]
